Move vscode-uitls/multi-file-system-watcher.ts to /common/vscode/multi-file-system-watcher.ts
