+++ extensions/ql-vscode/src/common/vscode/multi-file-system-watcher.ts
@@ -1,4 +1,4 @@
-import { DisposableObject } from "../pure/disposable-object";
+import { DisposableObject } from "../../pure/disposable-object";
 import { EventEmitter, Event, Uri, GlobPattern, workspace } from "vscode";
 
 /**
