+++ extensions/ql-vscode/src/qltest-discovery.ts
@@ -8,7 +8,7 @@ import {
   WorkspaceFolder,
   env,
 } from "vscode";
-import { MultiFileSystemWatcher } from "./vscode-utils/multi-file-system-watcher";
+import { MultiFileSystemWatcher } from "./common/vscode/multi-file-system-watcher";
 import { CodeQLCliServer } from "./codeql-cli/cli";
 import { pathExists } from "fs-extra";
